requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.23.5

# Database connections
sqlalchemy>=2.0.0
pyodbc>=4.0.0
duckdb>=0.9.0

# Google APIs
gspread>=5.0.0
gspread-dataframe>=3.3.0
oauth2client>=4.1.3
google-api-python-client>=2.0.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# Excel file handling
openpyxl>=3.1.0
xlsxwriter>=3.0.5